pyproject.toml
[project]
name = "img"
version = "0.1.0"
description = "Z-Image-Turbo REST Service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "python-dotenv>=1.0.0",
    "diffusers>=0.30.0",
    "torch>=2.0.0",
    "transformers>=4.40.0",
    "accelerate>=0.30.0",
    "pydantic>=2.0.0",
    "pillow>=10.0.0",
    "modelscope>=1.36.0",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
default = true
